requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==3.4.7
daphne==3.0.2
Django==3.1.6
django-cleanup==6.0.0
django-cors-headers==3.8.0
django-crispy-forms==1.11.2
django-embed-video==1.4.0
django-materializecss-form==1.1.17
django-rest-auth==0.9.5
django-rest-framework-base64-fields==0.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.1.1
drf-yasg==1.20.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
incremental==21.3.0
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
pika==1.2.0
Pillow==8.2.0
pipdeptree==2.2.0
plumbum==1.7.0
progressbar==2.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
Pyro4==4.80
python-telegram-bot==13.9
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==301
requests==2.26.0
rpyc==5.0.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
serpent==1.40
service-identity==21.1.0
simplejson==3.17.5
six==1.16.0
soaplib==1.0.0
sqlparse==0.4.1
suds-py3==1.4.4.1
tornado==6.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
zope.interface==5.4.0
